devide src
