+++ src/model_template.py
@@ -62,3 +62,31 @@ class PlotAndEvaluate(object):
                 columns=[kinds]
                 ).plot()
         self._save_png(save)
+
+
+class ModelTemplate(PlotAndEvaluate):
+    """
+    predict by linear model
+    """
+    def __init__(self, features, labels, terms):
+        """
+        predict and evaluate from features amd labels
+        :param features: dict
+        :param labels: dict
+        """
+        for asserted_value in [features, labels, terms]:
+            assert set(asserted_value.keys()) == {'train', 'valid', 'test'}
+        self.terms = terms
+        self.labels = labels
+        self.features = features
+        self.predicted_labels = None
+
+    def predict_all(self, verbose=None):
+        if verbose is None:
+            self.predicted_labels = {key: self.model.predict(
+                self.features[key]
+            ) for key in self.features.keys()}
+        else:
+            self.predicted_labels = {key: self.model.predict(
+                self.features[key], verbose=0
+            ) for key in self.features.keys()}
